+++ CHANGELOG.md
@@ -4,6 +4,7 @@ GAMSPy CHANGELOG
 GAMSPy 0.13.6
 -------------
 - General
+  - Make all file read and writes with utf-8 encoding.
   - Fix model instance record columns.
 - Testing
   - Add a test to verify the columns of symbols in model instance solves.